Continued refactoring
