+++ src/resource_manager/mod.rs
@@ -1,14 +1,19 @@
 mod commands;
-//TODO: Split this mod up
-mod resource_manager;
 mod hooks;
 mod bfresourcemgr;
+mod lazyresourcemap;
+mod ztfile;
+mod handlers;
+mod openzt_mods;
+mod legacy_loading;
+mod ztd;
 
 use commands::init_commands;
 use hooks::init_hooks;
 
-
-pub use resource_manager::{add_handler, modify_ztfile_as_animation, modify_ztfile_as_ini, Handler, RunStage, OPENZT_DIR0};
+pub use legacy_loading::OPENZT_DIR0;
+pub use ztfile::{modify_ztfile_as_animation, modify_ztfile_as_ini};
+pub use handlers::{add_handler, Handler, RunStage};
 
 pub fn init() {
     init_hooks();
