--- src/resource_manager/resource_manager.rs
@@ -1,1507 +1,0 @@
-use std::{fmt::Display, slice, str};
-use std::{
-    collections::{HashMap, HashSet}, ffi::CString, fmt, fs::File, io::{self, BufReader, Read}, path::{Path, PathBuf}, sync::{Mutex, Arc},
-};
-
-use anyhow::{anyhow, Context};
-
-use bf_configparser::ini::{Ini, WriteOptions};
-use once_cell::sync::Lazy;
-use tracing::{error, info};
-use walkdir::WalkDir;
-use zip::read::{ZipArchive, ZipFile};
-
-use regex::Regex;
-use crate::resource_manager::bfresourcemgr::BFResourcePtr;
-
-use crate::{
-    animation::Animation,
-    debug_dll::{get_from_memory, get_string_from_memory, save_to_memory},
-    mods,
-    string_registry::add_string_to_registry,
-};
-
-#[derive(Debug, Clone)]
-pub enum ZTFile {
-    Text(CString, ZTFileType, u32),
-    RawBytes(Box<[u8]>, ZTFileType, u32),
-}
-
-#[derive(Debug, Clone)]
-pub enum ZTFileType {
-    Ai,
-    Ani,
-    Cfg,
-    Lyt,
-    Scn,
-    Uca,
-    Ucs,
-    Ucb,
-    Ini,
-    Txt,
-    Toml,
-    Animation,
-    Palette,
-    Tga,
-    Wav,
-    Lle,
-    Bmp,
-    Zoo,
-}
-
-impl TryFrom<&Path> for ZTFileType {
-    type Error = &'static str;
-
-    fn try_from(path: &Path) -> Result<Self, Self::Error> {
-        let extension = path.extension().unwrap_or_default().to_ascii_lowercase();
-        Ok(match extension.to_str().unwrap_or_default() {
-            "ai" => ZTFileType::Ai,
-            "ani" => ZTFileType::Ani,
-            "cfg" => ZTFileType::Cfg,
-            "lyt" => ZTFileType::Lyt,
-            "scn" => ZTFileType::Scn,
-            "uca" => ZTFileType::Uca,
-            "ucs" => ZTFileType::Ucs,
-            "ucb" => ZTFileType::Ucb,
-            "ini" => ZTFileType::Ini,
-            "txt" => ZTFileType::Txt,
-            "toml" => ZTFileType::Toml,
-            "tga" => ZTFileType::Tga,
-            "wav" => ZTFileType::Wav,
-            "lle" => ZTFileType::Lle,
-            "bmp" => ZTFileType::Bmp,
-            "pal" => ZTFileType::Palette,
-            "zoo" => ZTFileType::Zoo,
-            "" => ZTFileType::Animation,
-            _ => return Err("Invalid file type"),
-        })
-    }
-}
-
-impl From<BFResourcePtr> for ZTFile {
-    fn from(bf_resource_ptr: BFResourcePtr) -> Self {
-        let filename = get_string_from_memory(bf_resource_ptr.bf_resource_name_ptr);
-        let file_extension = Path::new(&filename)
-            .extension()
-            .unwrap_or_default()
-            .to_str()
-            .unwrap_or_default();
-        let file_size = bf_resource_ptr.content_size;
-        let data = bf_resource_ptr.data_ptr;
-        match file_extension {
-            "ai" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Ai,
-                file_size,
-            ),
-            "cfg" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Cfg,
-                file_size,
-            ),
-            "lyt" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Lyt,
-                file_size,
-            ),
-            "scn" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Scn,
-                file_size,
-            ),
-            "uca" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Uca,
-                file_size,
-            ),
-            "ucs" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Ucs,
-                file_size,
-            ),
-            "ucb" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Ucb,
-                file_size,
-            ),
-            "ani" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Ani,
-                file_size,
-            ),
-            "ini" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Ini,
-                file_size,
-            ),
-            "txt" => ZTFile::Text(
-                unsafe { CString::from_raw(data as *mut i8) },
-                ZTFileType::Txt,
-                file_size,
-            ),
-            "tga" => ZTFile::RawBytes(
-                unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(data as *mut _, file_size as usize))
-                },
-                ZTFileType::Tga,
-                file_size,
-            ),
-            "pal" => ZTFile::RawBytes(
-                unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(data as *mut _, file_size as usize))
-                },
-                ZTFileType::Palette,
-                file_size,
-            ),
-            "wav" => ZTFile::RawBytes(
-                unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(data as *mut _, file_size as usize))
-                },
-                ZTFileType::Wav,
-                file_size,
-            ),
-            "lle" => ZTFile::RawBytes(
-                unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(data as *mut _, file_size as usize))
-                },
-                ZTFileType::Lle,
-                file_size,
-            ),
-            "bmp" => ZTFile::RawBytes(
-                unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(data as *mut _, file_size as usize))
-                },
-                ZTFileType::Bmp,
-                file_size,
-            ),
-            _ => ZTFile::RawBytes(
-                unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(data as *mut _, file_size as usize))
-                },
-                ZTFileType::Animation,
-                file_size,
-            ),
-        }
-    }
-}
-
-// TODO: These should probably be removed, lots of instances where you don't want to blindly get the ZTFileType from the file name
-impl ZTFile {
-    pub fn new(file_name: String, file_size: u32, data: Box<[u8]>) -> anyhow::Result<ZTFile> {
-        let file_extension = Path::new(&file_name).extension().unwrap_or_default().to_str().unwrap_or_default();
-        match file_extension {
-            "ai" | "cfg" | "lyt" | "scn" | "uca" | "ucs" | "ucb" | "ani" | "ini" | "txt" | "toml" => {
-
-                ZTFile::new_text(file_name, file_size, CString::new(data)?)
-            },
-            // Handles "tga", "pal", "wav", "lle", "bmp" or ""
-            _ => {
-                Ok(ZTFile::new_raw_bytes(file_name, file_size, data))
-            },
-        }
-    }
-
-    pub fn new_text_from_bytes(file_name: String, file_size: u32, data: Box<[u8]>) -> anyhow::Result<ZTFile> {
-        let c_string = CString::new(str::from_utf8(&data)?)?;
-        ZTFile::new_text(file_name, file_size, c_string)
-    }
-    
-    pub fn new_text(file_name: String, file_size: u32, data: CString) -> anyhow::Result<ZTFile> {
-        let file_extension = Path::new(&file_name).extension().unwrap_or_default().to_str().unwrap_or_default();
-        match file_extension {
-            "ai" => Ok(ZTFile::Text(data, ZTFileType::Ai, file_size)),
-            "cfg" => Ok(ZTFile::Text(data, ZTFileType::Cfg, file_size)),
-            "lyt" => Ok(ZTFile::Text(data, ZTFileType::Lyt, file_size)),
-            "scn" => Ok(ZTFile::Text(data, ZTFileType::Scn, file_size)),
-            "uca" => Ok(ZTFile::Text(data, ZTFileType::Uca, file_size)),
-            "ucs" => Ok(ZTFile::Text(data, ZTFileType::Ucs, file_size)),
-            "ucb" => Ok(ZTFile::Text(data, ZTFileType::Ucb, file_size)),
-            "ani" => Ok(ZTFile::Text(data, ZTFileType::Ani, file_size)),
-            "ini" => Ok(ZTFile::Text(data, ZTFileType::Ini, file_size)),
-            "txt" => Ok(ZTFile::Text(data, ZTFileType::Txt, file_size)),
-            "toml" => Ok(ZTFile::Text(data, ZTFileType::Toml, file_size)),
-            _ => Err(anyhow!("Invalid file type")),
-        }
-    }
-
-    pub fn new_raw_bytes(file_name: String, file_size: u32, data: Box<[u8]>) -> ZTFile {
-        let file_extension = Path::new(&file_name).extension().unwrap_or_default().to_str().unwrap_or_default();
-        match file_extension {
-            "tga" => ZTFile::RawBytes(data, ZTFileType::Tga, file_size),
-            "pal" => ZTFile::RawBytes(data, ZTFileType::Palette, file_size),
-            "wav" => ZTFile::RawBytes(data, ZTFileType::Wav, file_size),
-            "lle" => ZTFile::RawBytes(data, ZTFileType::Lle, file_size),
-            "bmp" => ZTFile::RawBytes(data, ZTFileType::Bmp, file_size),
-            _ => ZTFile::RawBytes(data, ZTFileType::Animation, file_size),
-        }
-    }
-}
-
-pub trait FromZipFile<T> {
-    fn from_zip_file(file: &mut ZipFile) -> io::Result<T>;
-}
-
-fn get_num_mod_ids() -> usize {
-    let binding = MOD_ID_SET.lock().unwrap();
-    binding.len()
-}
-
-fn ztfile_to_raw_resource(path: &str, file_name: String, ztfile: ZTFile)  -> anyhow::Result<u32> {
-    let mut ztd_path = path.to_string();
-    ztd_path = ztd_path.replace("./", "zip::./").replace('\\', "/");
-    let lowercase_filename = file_name.to_lowercase();
-
-    let bf_zip_name_ptr = match CString::new(ztd_path.clone()) {
-        Ok(c_string) => c_string.into_raw() as u32,
-        Err(e) => {
-            return Err(anyhow!("Error converting zip name to CString: {} -> {}", ztd_path, e));
-        }
-    };
-    let bf_resource_name_ptr = match CString::new(lowercase_filename.clone()) {
-        Ok(c_string) => c_string.into_raw() as u32,
-        Err(e) => {
-            return Err(anyhow!("Error converting resource name to CString: {} -> {}", lowercase_filename, e));
-        }
-    };
-
-    match ztfile {
-        ZTFile::Text(data, _, length) => {
-            let ptr = data.into_raw() as u32;
-            let resource_ptr = Box::into_raw(Box::new(BFResourcePtr {
-                num_refs: 100, // We set this very high to prevent the game from unloading the resource
-                bf_zip_name_ptr,
-                bf_resource_name_ptr,
-                data_ptr: ptr,
-                content_size: length,
-            }));
-
-            Ok(resource_ptr as _)
-        }
-        ZTFile::RawBytes(data, _, length) => {
-            let ptr = data.as_ptr() as u32;
-            std::mem::forget(data);
-            let resource_ptr = Box::into_raw(Box::new(BFResourcePtr {
-                num_refs: 100, // We set this very high to prevent the game from unloading the resource
-                bf_zip_name_ptr,
-                bf_resource_name_ptr,
-                data_ptr: ptr,
-                content_size: length,
-            }));
-
-            Ok(resource_ptr as _)
-        }
-    }
-}
-
-fn add_ztfile(path: &Path, file_name: String, ztfile: ZTFile) {
-    let Some(ztd_path) = path.to_str() else {
-        error!("Failed to convert path to string: {}", path.display());
-        return;
-    };
-    let mut ztd_path = ztd_path.to_string();
-    ztd_path = ztd_path.replace("./", "zip::./").replace('\\', "/");
-    let lowercase_filename = file_name.to_lowercase();
-
-    let mut binding = LAZY_RESOURCE_MAP.lock().unwrap();
-
-    let bf_zip_name_ptr = match CString::new(ztd_path.clone()) {
-        Ok(c_string) => c_string.into_raw() as u32,
-        Err(e) => {
-            error!("Error converting zip name to CString: {} -> {}", ztd_path, e);
-            return;
-        }
-    };
-    let bf_resource_name_ptr = match CString::new(lowercase_filename.clone()) {
-        Ok(c_string) => c_string.into_raw() as u32,
-        Err(e) => {
-            error!("Error converting resource name to CString: {} -> {}", lowercase_filename, e);
-            return;
-        }
-    };
-
-    match ztfile {
-        ZTFile::Text(data, file_type, length) => {
-            let ptr = data.into_raw() as u32;
-            let resource_ptr = Box::into_raw(Box::new(BFResourcePtr {
-                num_refs: 100, // We set this very high to prevent the game from unloading the resource
-                bf_zip_name_ptr,
-                bf_resource_name_ptr,
-                data_ptr: ptr,
-                content_size: length,
-            }));
-
-            binding.insert_custom(file_name.clone(), file_type, resource_ptr as u32);
-        }
-        ZTFile::RawBytes(data, file_type, length) => {
-            let ptr = data.as_ptr() as u32;
-            std::mem::forget(data);
-            let resource_ptr = Box::into_raw(Box::new(BFResourcePtr {
-                num_refs: 100, // We set this very high to prevent the game from unloading the resource
-                bf_zip_name_ptr,
-                bf_resource_name_ptr,
-                data_ptr: ptr,
-                content_size: length,
-            }));
-
-            binding.insert_custom(lowercase_filename.clone(), file_type, resource_ptr as u32);
-        }
-    }
-}
-
-pub fn modify_ztfile<F>(file_name: &str, modifier: F) -> Result<(), &'static str>
-where
-    F: Fn(&mut BFResourcePtr),
-{
-    let Some(bf_resource_ptr_ptr) = get_file_ptr(file_name) else {
-        info!("File not found: {}", file_name);
-        return Err("File not found");
-    };
-    let mut bf_resource_ptr = get_from_memory::<BFResourcePtr>(bf_resource_ptr_ptr);
-
-    modifier(&mut bf_resource_ptr);
-
-    save_to_memory::<BFResourcePtr>(bf_resource_ptr_ptr, bf_resource_ptr.clone());
-
-    Ok(())
-}
-
-pub fn modify_ztfile_as_ini<F>(file_name: &str, modifier: F) -> Result<(), &'static str>
-where
-    F: Fn(&mut Ini),
-{
-    modify_ztfile(file_name, |file: &mut BFResourcePtr| {
-        let c_string = unsafe { CString::from_raw(file.data_ptr as *mut i8) };
-        let c_string_as_string = c_string.to_string_lossy().to_string();
-        let mut cfg = Ini::new_cs();
-        cfg.set_comment_symbols(&[';', '#', ':']);
-        if let Err(err) = cfg.read(c_string_as_string) {
-            error!("Error reading ini: {}", err);
-            return;
-        };
-
-        modifier(&mut cfg);
-
-        let mut write_options = WriteOptions::default();
-        write_options.space_around_delimiters = true;
-        write_options.blank_lines_between_sections = 1;
-        let new_string = cfg.pretty_writes(&write_options);
-        file.content_size = new_string.len() as u32;
-
-        let Ok(new_c_string) = CString::new(new_string) else {
-            error!("Error converting ini to CString after modifying {} writing unchanged version", file_name);
-            return;
-        };
-        file.data_ptr = new_c_string.into_raw() as u32;
-    })
-}
-
-pub fn modify_ztfile_as_animation<F>(file_name: &str, modifier: F) -> Result<(), &'static str>
-where
-    F: Fn(&mut Animation),
-{
-    modify_ztfile(file_name, |file: &mut BFResourcePtr| {
-        info!("Modifying animation");
-        let data_vec: Box<[u8]> = unsafe {
-            Box::from_raw(slice::from_raw_parts_mut(
-                file.data_ptr as *mut _,
-                file.content_size as usize,
-            ))
-        };
-        let mut animation = Animation::parse(&data_vec);
-        modifier(&mut animation);
-        let (new_animation_bytes, length) = animation.write();
-        let boxed_slice = new_animation_bytes.into_boxed_slice();
-        let data_ptr = boxed_slice.as_ptr() as u32;
-        std::mem::forget(boxed_slice);
-        file.data_ptr = data_ptr;
-        file.content_size = length as u32;
-    })
-}
-
-pub const OPENZT_DIR0: &str = "openzt_resource";
-
-///Indicates when the handler should be called
-/// BeforeOpenZTMods means they are run on any files in but before any OpenZT mods are loaded
-/// AfterOpenZTMods is the same as above but after OpenZT mods are loaded
-/// AfterFiltering is run on files that are referenced in *.cfg files only
-#[derive(Clone, PartialEq)]
-pub enum RunStage {
-    BeforeOpenZTMods,
-    AfterOpenZTMods,
-    AfterFiltering
-}
-
-pub type IniHandlerFunction = fn(&str, &str, Ini) -> Option<(String, String, Ini)>;
-pub type AnimationHandlerFunction = fn(&str, &str, Animation) -> Option<(String, String, Animation)>;
-pub type RawBytesHandlerFunction = fn(&str, &str, Box<[u8]>) -> Option<(String, String, Box<[u8]>)>;
-
-pub struct HandlerBuilder<const HAS_STAGE: bool, const HAS_HANDLER: bool> {
-    matcher_prefix: Option<String>,
-    matcher_suffix: Option<String>,
-    ini_handler: Option<IniHandlerFunction>,
-    animation_handler: Option<AnimationHandlerFunction>,
-    raw_bytes_handler: Option<RawBytesHandlerFunction>,
-    stage: Option<RunStage>,
-}
-
-impl<const HAS_STAGE: bool, const HAS_HANDLER: bool> HandlerBuilder<HAS_STAGE, HAS_HANDLER> {
-    pub fn prefix(mut self, prefix: &str) -> HandlerBuilder<HAS_STAGE, HAS_HANDLER> {
-        self.matcher_prefix = Some(prefix.to_owned());
-        self
-    }
-
-    pub fn suffix(mut self, suffix: &str) -> HandlerBuilder<HAS_STAGE, HAS_HANDLER> {
-        self.matcher_suffix = Some(suffix.to_owned());
-        self
-    }
-
-    pub fn ini_handler(self, handler: IniHandlerFunction) -> HandlerBuilder<HAS_STAGE, true> {
-        HandlerBuilder {
-            matcher_prefix: self.matcher_prefix,
-            matcher_suffix: self.matcher_suffix,
-            ini_handler: Some(handler),
-            animation_handler: self.animation_handler,
-            raw_bytes_handler: self.raw_bytes_handler,
-            stage: self.stage
-        }
-    }
-
-    pub fn animation_handler(self, handler: AnimationHandlerFunction) -> HandlerBuilder<HAS_STAGE, true> {
-        HandlerBuilder {
-            matcher_prefix: self.matcher_prefix,
-            matcher_suffix: self.matcher_suffix,
-            ini_handler: self.ini_handler,
-            animation_handler: Some(handler),
-            raw_bytes_handler: self.raw_bytes_handler,
-            stage: self.stage
-        }
-    }
-
-    pub fn raw_bytes_handler(self, handler: RawBytesHandlerFunction) -> HandlerBuilder<HAS_STAGE, true> {
-        HandlerBuilder {
-            matcher_prefix: self.matcher_prefix,
-            matcher_suffix: self.matcher_suffix,
-            ini_handler: self.ini_handler,
-            animation_handler: self.animation_handler,
-            raw_bytes_handler: Some(handler),
-            stage: self.stage
-        }
-    }
-
-    pub fn run_stage(self, stage: RunStage) -> HandlerBuilder<true, HAS_HANDLER> {
-        HandlerBuilder {
-            matcher_prefix: self.matcher_prefix,
-            matcher_suffix: self.matcher_suffix,
-            ini_handler: self.ini_handler,
-            animation_handler: self.animation_handler,
-            raw_bytes_handler: self.raw_bytes_handler,
-            stage: Some(stage)
-        }
-    }
-}
-
-impl HandlerBuilder<true, true> {
-    pub fn build(self) -> Handler {
-        unsafe {
-            Handler {
-                matcher_prefix: self.matcher_prefix,
-                matcher_suffix: self.matcher_suffix,
-                ini_handler: self.ini_handler,
-                animation_handler: self.animation_handler,
-                raw_bytes_handler: self.raw_bytes_handler,
-                stage: self.stage.unwrap_unchecked()
-            }
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct Handler {
-    matcher_prefix: Option<String>,
-    matcher_suffix: Option<String>,
-    ini_handler: Option<IniHandlerFunction>,
-    animation_handler: Option<AnimationHandlerFunction>,
-    raw_bytes_handler: Option<RawBytesHandlerFunction>,
-    stage: RunStage
-}
-
-impl Handler {
-    pub fn builder() -> HandlerBuilder<false, false> {
-        HandlerBuilder::<false, false> {
-            matcher_prefix: None,
-            matcher_suffix: None,
-            ini_handler: None,
-            animation_handler: None,
-            raw_bytes_handler: None,
-            stage: None
-        }
-    }
-
-    fn handle(&self, file_name: &String) {
-        if let Some(prefix) = &self.matcher_prefix {
-            if !file_name.starts_with(prefix) {
-                return;
-            }
-        }
-        if let Some(suffix) = &self.matcher_suffix {
-            if !file_name.ends_with(suffix) {
-                return;
-            }
-        }
-
-        let file_type = match ZTFileType::try_from(Path::new(&file_name)) {
-            Ok(file_type) => file_type,
-            Err(e) => {
-                error!("Error getting file type: {} error: {}", file_name, e);
-                return;
-            }
-        };
-
-        let new_file = match file_type {
-            ZTFileType::Ini | ZTFileType::Ai | ZTFileType::Ani | ZTFileType::Cfg | ZTFileType::Lyt | ZTFileType::Scn | ZTFileType::Uca | ZTFileType::Ucs | ZTFileType::Ucb => {
-                if let Some(handler) = self.ini_handler {
-                    info!("Ini Handler {} {} is handling file: {}", self.matcher_prefix.as_deref().unwrap_or_default(), self.matcher_suffix.as_deref().unwrap_or_default(), file_name);
-                    let Some((archive_name, file)) = get_file(file_name) else {
-                        error!("Error getting file: {}", file_name);
-                        return;
-                    };
-                    let mut ini = Ini::new_cs();
-                    ini.set_comment_symbols(&[';', '#', ':']);
-
-                    let Ok(input_string) = str::from_utf8(&file) else {
-                        error!("Error converting file to string: {}", file_name);
-                        return;
-                    };
-                    if let Err(e) = ini.read(input_string.to_string()) {
-                        error!("Error reading ini {}: {}", file_name, e);
-                        return;
-                    }
-                    if let Some((new_archive_name, new_file_path, new_ini)) = handler(&archive_name, file_name, ini) {
-                        let mut write_options = WriteOptions::default();
-                        write_options.space_around_delimiters = true;
-                        write_options.blank_lines_between_sections = 1;
-                        let new_string = new_ini.pretty_writes(&write_options);
-                        let new_string_length = new_string.len() as u32;
-
-                        match CString::new(new_string) {
-                            Ok(new_c_string) => {
-                                Some((new_archive_name, new_file_path, ZTFile::Text(new_c_string, file_type, new_string_length)))
-                            },
-                            Err(_) => {
-                                error!("Error converting ini to CString after modifying {}", file_name);
-                                None
-                            }
-                        }
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                }
-            }
-            ZTFileType::Animation => {
-                if let Some(handler) = self.animation_handler {
-                    info!("Animation Handler {} {} is handling file: {}", self.matcher_prefix.as_deref().unwrap_or_default(), self.matcher_suffix.as_deref().unwrap_or_default(), file_name);
-                    let Some((archive_name, file)) = get_file(file_name) else {
-                        error!("Error getting file: {}", file_name);
-                        return;
-                    };
-                    let animation = Animation::parse(&file);
-                    if let Some((new_archive_name, new_file_path, new_animation)) = handler(&archive_name, file_name, animation) {
-                        let (new_animation_bytes, animation_size) = new_animation.write();
-                        Some((new_archive_name, new_file_path, ZTFile::RawBytes(new_animation_bytes.into_boxed_slice(), ZTFileType::Animation, animation_size as u32)))
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                }
-            }
-            ZTFileType::Bmp | ZTFileType::Lle | ZTFileType::Tga | ZTFileType::Wav | ZTFileType::Palette => {
-                if let Some(handler) = self.raw_bytes_handler {
-                    info!("Raw Bytes Handler {} {} is handling file: {}", self.matcher_prefix.as_deref().unwrap_or_default(), self.matcher_suffix.as_deref().unwrap_or_default(), file_name);
-                    let Some((archive_name, file)) = get_file(file_name) else {
-                        error!("Error getting file: {}", file_name);
-                        return;
-                    };
-                    if let Some((new_archive_name, new_file_path, new_data)) = handler(&archive_name, file_name, file) {
-                        let new_data_len = new_data.len() as u32;
-                        Some((new_archive_name, new_file_path, ZTFile::RawBytes(new_data, file_type, new_data_len)))
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                }
-            }
-
-            _ => return
-        };
-
-        if let Some((new_archive_name, new_file_name, ztfile)) = new_file {
-            add_ztfile(Path::new(&new_archive_name), new_file_name, ztfile)
-        }
-
-    }
-}
-
-fn get_file(file_name: &str) -> Option<(String, Box<[u8]>)> {
-    let mut map = LAZY_RESOURCE_MAP.lock().unwrap();
-    match map.get(file_name) {
-        Ok(Some(file)) => {
-            let resource_ptr = get_from_memory::<BFResourcePtr>(file.data);
-            let tmp_slice = unsafe {slice::from_raw_parts(resource_ptr.data_ptr as *const _, resource_ptr.content_size as usize) };
-            let mut new_slice = vec![0; resource_ptr.content_size as usize];
-            new_slice.copy_from_slice(tmp_slice);
-            Some((get_string_from_memory(resource_ptr.bf_zip_name_ptr), new_slice.into_boxed_slice()))
-        },
-        Ok(None) => {
-            info!("File not found: {}", file_name);
-            None
-        },
-        Err(e) => {
-            info!("Error getting file: {} error: {}", file_name, e);
-            None
-        }
-    }
-}
-
-// Note: We are excluding ztat* files until we need to override anything inside them, as they have a rediculous amount of files
-fn get_ztd_resources(dir: &Path, recursive: bool) -> Vec<PathBuf> {
-    let mut resources = Vec::new();
-    if !dir.is_dir() {
-        return resources;
-    }
-    let walker = WalkDir::new(dir).follow_links(true).max_depth(if recursive { 0 } else { 1 });
-    for entry in walker {
-        let entry = match entry {
-            Ok(entry) => entry,
-            Err(e) => {
-                error!("Error walking directory: {}", e);
-                continue;
-            }
-        };
-        let Some(filename) = entry.file_name().to_str() else {
-            error!("Error getting filename: {:?}", entry);
-            continue;
-        };
-        if filename.to_lowercase().ends_with(".ztd") && !filename.starts_with("ztat") {
-            resources.push(entry.path().to_path_buf());
-        }
-    }
-    resources
-}
-
-pub static LAZY_RESOURCE_MAP: Lazy<Mutex<LazyResourceMap>> = Lazy::new(|| Mutex::new(LazyResourceMap::new()));
-
-pub struct LazyResourceMap {
-    map: HashMap<String, LazyResource>,
-}
-
-#[derive(Clone)]
-enum ResourceBacking {
-    LazyZipFile{archive_name: String, archive: Arc<Mutex<ZipArchive<BufReader<File>>>>},
-    LoadedZipFile{archive_name: String, archive: Arc<Mutex<ZipArchive<BufReader<File>>>>, data: u32},
-    Custom{data: u32},
-}
-
-struct ConcreteResource {
-    archive_name: Option<String>,
-    filename: String,
-    type_: ZTFileType,
-    data: u32,
-}
-
-struct LazyResource {
-    pub backing: ResourceBacking,
-    pub filename: String,
-    pub type_: ZTFileType,
-}
-
-impl LazyResourceMap {
-    fn new() -> Self {
-        Self {
-            map: HashMap::new(),
-        }
-    }
-
-    fn _drop(&mut self, file_name: String) -> Option<()> {
-        let value = self.map.remove(&file_name)?;
-        
-        self.drop_inner(value);
-        Some(())
-    }
-
-    fn drop_inner(&mut self, resource: LazyResource) {
-        let data = match resource.backing {
-            ResourceBacking::LoadedZipFile{data, archive_name: _, archive: _ } => {
-                data
-            },
-            ResourceBacking::Custom{data} => {
-                data
-            },
-            ResourceBacking::LazyZipFile{archive_name: _, archive: _} => {
-                return;
-            }
-        };
-        let bf_resource_ptr = unsafe { Box::from_raw(data as *mut BFResourcePtr) };
-        match resource.type_ {
-            ZTFileType::Ini | ZTFileType::Ai | ZTFileType::Ani | ZTFileType::Cfg | ZTFileType::Lyt | ZTFileType::Scn | ZTFileType::Uca | ZTFileType::Ucs | ZTFileType::Ucb | ZTFileType::Toml | ZTFileType::Txt => {
-                let data_string = unsafe { CString::from_raw(data as *mut i8) };
-                drop(data_string);
-            }
-            ZTFileType::Animation | ZTFileType::Bmp | ZTFileType::Lle | ZTFileType::Tga | ZTFileType::Wav | ZTFileType::Palette | ZTFileType::Zoo => {
-                let data_vec: Box<[u8]> = unsafe {
-                    Box::from_raw(slice::from_raw_parts_mut(
-                        bf_resource_ptr.data_ptr as *mut _,
-                        bf_resource_ptr.content_size as usize,
-                    ))
-                };
-                drop(data_vec);
-            }
-        }
-        drop(bf_resource_ptr);
-    }
-    
-    fn insert_lazy(&mut self, archive_path: String, file_name: String, archive: Arc<Mutex<ZipArchive<BufReader<File>>>>) {
-        let file_type = match ZTFileType::try_from(Path::new(&file_name)) {
-            Ok(file_type) => file_type,
-            Err(e) => {
-                error!("Error inserting file: {} error: {}", file_name, e);
-                return;
-            }
-        };
-
-        if let Some(existing) = self.map.insert(file_name.clone().to_ascii_lowercase(), LazyResource {
-            backing: ResourceBacking::LazyZipFile{archive_name: archive_path.clone(), archive},
-            filename: file_name.clone(),
-            type_: file_type,
-        }) {
-            self.drop_inner(existing);
-        }
-    }
-
-    fn insert_loaded(&mut self, resource: LazyResource) {
-        if let Some(existing) = self.map.insert(resource.filename.to_ascii_lowercase(), resource) {
-            self.drop_inner(existing);
-        }
-    }
-
-    fn insert_custom(&mut self, file_name: String, file_type: ZTFileType, data: u32) {
-        if let Some(existing) = self.map.insert(file_name.to_ascii_lowercase(), LazyResource {
-            backing: ResourceBacking::Custom{data},
-            filename: file_name.clone(),
-            type_: file_type,
-        }) {
-            self.drop_inner(existing);
-        }
-    }
-
-    fn get(&mut self, key: &str) -> anyhow::Result<Option<ConcreteResource>> {
-        let lowercase_key = key.to_ascii_lowercase();
-        let Some(resource) = self.map.get_mut(&lowercase_key) else {
-            info!("LazyResource not found: {}", lowercase_key);
-            return Ok(None);
-        };
-
-        // TODO: Use std::mem::take/replace to avoid cloning
-        let (archive_name, data) = match resource.backing.clone() {
-            ResourceBacking::LazyZipFile{archive_name, archive} => {
-                let mut binding = archive.lock().unwrap();
-                let mut file = binding.by_name(&resource.filename).with_context(|| format!("Error finding file in archive: {}", resource.filename))?;
-                let mut file_buffer = vec![0u8; file.size() as usize].into_boxed_slice();
-
-                file.read_exact(&mut file_buffer).with_context(|| format!("Error reading file: {}", resource.filename))?;
-                let ztfile = ZTFile::new(resource.filename.clone(), file_buffer.len() as u32, file_buffer)?;
-                let data = ztfile_to_raw_resource(&archive_name, resource.filename.clone(), ztfile)?;
-                resource.backing = ResourceBacking::LoadedZipFile{archive_name: archive_name.clone(), archive: archive.clone(), data};
-                (Some(archive_name), data)
-            },
-            ResourceBacking::LoadedZipFile{archive_name, archive: _, data} => {
-                (Some(archive_name), data)
-            },
-            ResourceBacking::Custom{data} => {
-                (None, data)
-            }
-        };
-
-        Ok(Some(ConcreteResource{
-            archive_name: archive_name.clone(),
-            filename: resource.filename.clone(),
-            type_: resource.type_.clone(),
-            data,
-        }))
-    }
-
-    fn loaded_len(&self) -> usize {
-        self.map.values().filter(|x| matches!(x.backing, ResourceBacking::LoadedZipFile{..} | ResourceBacking::Custom{..})).count()
-    }
-
-    fn not_loaded_len(&self) -> usize {
-        self.map.values().filter(|x| matches!(x.backing, ResourceBacking::LazyZipFile{..})).count()
-    }
-
-    fn len(&self) -> usize {
-        self.map.len()
-    }
-
-    fn contains_key(&self, key: &str) -> bool {
-        self.map.contains_key(key)
-    }
-
-    pub fn files(&self) -> Box<dyn Iterator<Item = String> + '_> {
-        Box::new(self.map.keys().cloned())
-    }
-}
-
-pub fn check_file(file_name: &str) -> bool {
-    let binding = LAZY_RESOURCE_MAP.lock().unwrap();
-    binding.contains_key(&file_name.to_lowercase())
-}
-
-pub fn get_file_ptr(file_name: &str) -> Option<u32> {
-    let mut binding = LAZY_RESOURCE_MAP.lock().unwrap();
-    if let Ok(Some(resource)) = binding.get(&file_name.to_lowercase()) {
-        Some(resource.data)
-    } else {
-        None
-    }
-}
-
-pub fn load_resources(paths: Vec<String>) {
-    use std::time::Instant;
-    let now = Instant::now();
-    let mut resource_count = 0;
-
-    paths.iter().rev().for_each(|path| {
-        let resources = get_ztd_resources(Path::new(path), false);
-        resources.iter().for_each(|resource| {
-            info!("Loading resource: {}", resource.display());
-            let file_name = resource.to_str().unwrap_or_default().to_lowercase();
-            if file_name.ends_with(".ztd") {
-                match handle_ztd(resource) {
-                    Ok(count) => resource_count += count,
-                    Err(err) => error!("Error loading ztd: {} -> {}", file_name, err),
-                }
-            }
-        });
-    });
-
-    let files = {
-        let map = LAZY_RESOURCE_MAP.lock().unwrap();
-
-        let elapsed = now.elapsed();
-        info!(
-            "Loaded {} mods and {} ({}) resources in: {:.2?}",
-            get_num_mod_ids(),
-            map.len(),
-            resource_count,
-            elapsed
-        );
-
-        map.files().collect::<Vec<String>>().into_iter()
-    };
-    
-
-    let now = Instant::now();
-
-    let data_mutex = RESOURCE_HANDLER_ARRAY.lock().unwrap();
-
-
-    info!("Running BeforeOpenZTMods handlers");
-    for handler in data_mutex.iter() {
-        if handler.stage == RunStage::BeforeOpenZTMods {
-            files.clone().for_each(|file| {
-                handler.handle(&file);
-            });
-        }
-    }
-
-    // TODO: Implement patching
-    // apply_patches();
-
-    info!("Running AfterOpenZTMods handlers");
-    for handler in data_mutex.iter() {
-        if handler.stage == RunStage::AfterOpenZTMods {
-            files.clone().for_each(|file| {
-                handler.handle(&file);
-            });
-        }
-    }
-
-    let mut filtered_files = Vec::new();
-
-    files.clone().for_each(|file| {
-        let extension = Path::new(&file).extension().unwrap_or_default().to_ascii_lowercase();
-        match extension.to_str().unwrap_or_default() {
-            "uca" | "ucs" | "ucb" => filtered_files.push(file),
-            "cfg" => {
-                let inner_filtered = parse_cfg(&file);
-                filtered_files.extend(inner_filtered);
-            }
-            _ => {}
-        }
-    });
-
-    info!("Loaded {} filtered files", filtered_files.len());
-
-    info!("Running AfterFiltering handlers");
-    for handler in data_mutex.iter() {
-        if handler.stage == RunStage::AfterFiltering {
-            filtered_files.clone().into_iter().for_each(|file| {
-               handler.handle(&file);
-            });
-        }
-    }
-    
-    let elapsed = now.elapsed();
-    info!(
-        "Extra handling took an extra: {:.2?}",
-        elapsed
-    );
-}
-
-fn handle_ztd(resource: &PathBuf) -> anyhow::Result<i32> {
-    let mut load_count = 0;
-    let file = File::open(resource).with_context(|| format!("Error opening file: {}", resource.display()))?;
-
-    let resource_string = resource
-        .clone()
-        .into_os_string()
-        .into_string()
-        .map_err(|e| anyhow::anyhow!("error converting resource path to string: {}", e.to_string_lossy()))?;
-
-    let buf_reader = BufReader::new(file);
-
-    let mut zip = zip::ZipArchive::new(buf_reader).with_context(|| format!("Error reading zip: {}", resource.display()))?;
-
-    let ztd_type = load_open_zt_mod(&mut zip)?;
-
-    if ztd_type == mods::ZtdType::Openzt {
-        return Ok(0);
-    }
-
-    let archive = Arc::new(Mutex::new(zip));
-
-    let mut map = LAZY_RESOURCE_MAP.lock().unwrap();
-    
-    archive.lock().unwrap().file_names().filter(|s| !s.ends_with("/")).for_each(|file_name| {
-        map.insert_lazy(resource_string.clone(), file_name.to_string(), archive.clone());
-        load_count += 1;
-    });
-
-    Ok(load_count)
-}
-
-
-fn parse_cfg(file_name: &String) -> Vec<String> {
-    if let Some(legacy_cfg) = get_legacy_cfg_type(file_name) {
-        info!("Legacy cfg: {} {:?}", file_name, legacy_cfg.cfg_type);
-
-        let Some((_archive_name, file)) = get_file(file_name) else {
-            error!("Error getting file: {}", file_name);
-            return Vec::new();
-        };
-        let mut ini = Ini::new_cs();
-        ini.set_comment_symbols(&[';', '#', ':']);
-        let Ok(input_string) = str::from_utf8(&file) else {
-            error!("Error converting file to string: {}", file_name);
-            return Vec::new();
-        };
-        if let Err(e) = ini.read(input_string.to_string()) {
-            error!("Error reading ini {}: {}", file_name, e);
-            return Vec::new();
-        }
-
-        match legacy_cfg.cfg_type {
-            LegacyCfgType::Ambient => parse_simple_cfg(&ini, "ambient"),
-            LegacyCfgType::Animal => parse_simple_cfg(&ini, "animals"), //parse_subtypes_cfg(&ini, "animals"),
-            LegacyCfgType::Building => parse_simple_cfg(&ini, "building"),
-            LegacyCfgType::Fence => parse_simple_cfg(&ini, "fences"), //parse_subtypes_cfg(&ini, "fences"),
-            LegacyCfgType::Filter => parse_simple_cfg(&ini, "filter"), //parse_subtypes_cfg(&ini, "filter"),
-            LegacyCfgType::Food => parse_simple_cfg(&ini, "food"),
-            LegacyCfgType::Free => parse_simple_cfg(&ini, "freeform"),
-            // LegacyCfgType::Fringe => Vec::new(),
-            LegacyCfgType::Guest => parse_simple_cfg(&ini, "guest"),
-            // LegacyCfgType::Help => Vec::new(),
-            LegacyCfgType::Item => parse_simple_cfg(&ini, "items"),
-            LegacyCfgType::Path => parse_simple_cfg(&ini, "paths"),
-            LegacyCfgType::Rubble => parse_simple_cfg(&ini, "other"),
-            // LegacyCfgType::Scenario => Vec::new(),
-            LegacyCfgType::Scenery => {
-                let mut results = parse_simple_cfg(&ini, "objects");
-                results.append(&mut parse_simple_cfg(&ini, "foliage"));
-                results.append(&mut parse_simple_cfg(&ini, "other"));
-                results
-            },
-            LegacyCfgType::Staff => parse_simple_cfg(&ini, "staff"), //parse_subtypes_cfg(&ini, "staff"),
-            LegacyCfgType::Tile => Vec::new(),
-            LegacyCfgType::Wall => parse_simple_cfg(&ini, "tankwall"), //parse_subtypes_cfg(&ini, "tankwall"),
-            // LegacyCfgType::Expansion => Vec::new(),
-            // LegacyCfgType::Show => Vec::new(),
-            // LegacyCfgType::Tank => Vec::new(),
-            // LegacyCfgType::UIInfoImage => Vec::new(),
-            // LegacyCfgType::Economy => Vec::new(),
-            _ => Vec::new(),
-        }
-    } else {
-        Vec::new()
-    }
-}
-
-fn parse_simple_cfg(file: &Ini, section_name: &str) -> Vec<String> {
-    let mut results = Vec::new();
-    if let Some(section) = file.get_map().unwrap_or_default().get(section_name) {
-        for (_, value) in section.iter() {
-            if let Some(value) = value {
-                if value.len() == 1 {
-                    results.push(value[0].clone());
-                }
-            };
-        }
-    }
-    results
-}
-
-#[derive(Debug)]
-enum LegacyCfgType {
-    Ambient,
-    Animal,
-    Building,
-    Fence,
-    Filter,
-    Food,
-    Free,
-    Fringe,
-    Guest,
-    Help,
-    Item,
-    Path,
-    Rubble,
-    Scenario,
-    Scenery,
-    Staff,
-    Tile,
-    Wall,
-    Expansion,
-    Show,
-    Tank,
-    UIInfoImage,
-    Economy,
-}
-
-#[derive(Debug)]
-struct LegacyCfg {
-    cfg_type: LegacyCfgType,
-    file_name: String,
-}
-
-fn map_legacy_cfg_type(file_type_str: &str, file_name: String) -> Result<LegacyCfg, String> {
-    match file_type_str {
-        "ambient" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Ambient,
-            file_name,
-        }),
-        "animal" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Animal,
-            file_name,
-        }),
-        "bldg" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Building,
-            file_name,
-        }),
-        "fences" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Fence,
-            file_name,
-        }),
-        "filter" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Filter,
-            file_name,
-        }),
-        "food" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Food,
-            file_name,
-        }),
-        "free" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Free,
-            file_name,
-        }),
-        "fringe" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Fringe,
-            file_name,
-        }),
-        "guests" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Guest,
-            file_name,
-        }),
-        "help" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Help,
-            file_name,
-        }),
-        "items" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Item,
-            file_name,
-        }),
-        "paths" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Path,
-            file_name,
-        }),
-        "rubble" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Rubble,
-            file_name,
-        }),
-        "scenar" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Scenario,
-            file_name,
-        }),
-        "scener" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Scenery,
-            file_name,
-        }),
-        "staff" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Staff,
-            file_name,
-        }),
-        "tile" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Tile,
-            file_name,
-        }),
-        "twall" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Wall,
-            file_name,
-        }),
-        "xpac" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Expansion,
-            file_name,
-        }),
-        "shows" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Show,
-            file_name,
-        }),
-        "tanks" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Tank,
-            file_name,
-        }),
-        "ui/infoimg" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::UIInfoImage,
-            file_name,
-        }),
-        "economy" => Ok(LegacyCfg {
-            cfg_type: LegacyCfgType::Economy,
-            file_name,
-        }),
-        _ => Err(format!("Unknown legacy cfg type: {}", file_type_str)),
-    }
-}
-
-static LEGACY_CFG_REGEX: Lazy<Regex> = Lazy::new(|| {
-    Regex::new(r"^((ambient|animal|bldg|fences|filter|food|free|fringe|guests|help|items|paths|rubble|scenar|scener|staff|tile|twall|xpac)[\w\-. ]*?\.cfg)|((shows|tanks|ui\/infoimg|economy)\.cfg)$")
-        .unwrap()
-});
-
-fn get_legacy_cfg_type(file_name: &str) -> Option<LegacyCfg> {
-    let capture = LEGACY_CFG_REGEX.captures(file_name)?;
-    match capture.iter().collect::<Vec<_>>().as_slice() {
-        [_, Some(file_name), Some(file_type), None, None] => {
-            map_legacy_cfg_type(file_type.as_str(), file_name.as_str().to_string()).ok()
-        }
-        [_, None, None, Some(file_name), Some(file_type)] => {
-            map_legacy_cfg_type(file_type.as_str(), file_name.as_str().to_string()).ok()
-        }
-        _ => {
-            None
-        }
-    }
-}
-
-// TODO: Make some ZipFile and ZipArchive wrappers and add these functions to them
-fn read_file_from_zip(zip: &mut ZipArchive<BufReader<File>>, file_name: &str) -> anyhow::Result<Box<[u8]>> {
-    let mut file = zip.by_name(file_name).with_context(|| format!("Error finding file in archive: {}", file_name))?;
-
-    let mut file_buffer = vec![0u8; file.size() as usize].into_boxed_slice();
-
-    file.read_exact(&mut file_buffer).with_context(|| format!("Error reading file: {}", file_name))?;
-
-    Ok(file_buffer)
-}
-
-fn read_file_from_zip_to_string(zip: &mut ZipArchive<BufReader<File>>, file_name: &str) -> anyhow::Result<String> {
-    let buffer = read_file_from_zip(zip, file_name)?;
-
-    Ok(str::from_utf8(&buffer)
-        .with_context(|| format!("Error converting file {} to utf8", file_name))?
-        .to_string())
-}
-
-fn zip_file_to_string(mut zip: ZipFile) -> anyhow::Result<String> {
-    let mut buffer = vec![0u8; zip.size() as usize].into_boxed_slice();
-    zip.read_exact(&mut buffer).with_context(|| format!("Error reading file: {}", zip.name()))?;
-
-    Ok(str::from_utf8(&buffer)
-        .with_context(|| format!("Error converting file {} to utf8", zip.name()))?
-        .to_string())
-}
-
-fn load_open_zt_mod(archive: &mut ZipArchive<BufReader<File>>) -> anyhow::Result<mods::ZtdType> {
-    if archive.by_name("meta.toml").is_err() {
-        return Ok(mods::ZtdType::Legacy);
-    }
-
-    let meta = toml::from_str::<mods::Meta>(&read_file_from_zip_to_string(archive, "meta.toml")?).with_context(|| "Failed to parse meta.toml")?;
-
-    if meta.ztd_type() == &mods::ZtdType::Legacy {
-        return Ok(mods::ZtdType::Legacy);
-    }
-
-    let mod_id = meta.mod_id().to_string();
-
-    if !add_new_mod_id(&mod_id) {
-        return Err(anyhow!("Mod already loaded: {}", mod_id));
-    }
-
-    info!("Loading OpenZT mod: {} {}", meta.name(), meta.mod_id());
-
-    let mut file_map: HashMap<String, Box<[u8]>> = HashMap::new();
-
-    for i in 0..archive.len() {
-        let mut file = archive
-            .by_index(i)
-            // TODO: Create type that wraps ZipArchive and provide archive name for better error reporting
-            .with_context(|| format!("Error reading zip file at index {}", i))?;
-
-        if file.is_dir() {
-            continue;
-        }
-        let file_name = file.name().to_string();
-
-        let mut file_buffer = vec![0; file.size() as usize].into_boxed_slice();
-        file.read_exact(&mut file_buffer).with_context(|| format!("Error reading file: {}", file_name))?;
-
-        file_map.insert(file_name, file_buffer);
-    }
-
-    let keys = file_map.keys().clone();
-
-    for file_name in keys {
-        if file_name.starts_with("defs/") {
-            load_def(&mod_id, file_name, &file_map)?;
-        }
-    }
-
-    Ok(meta.ztd_type().clone())
-}
-
-// Map between the id ZT uses to reference locations/habitats and the string ptr of the animation (icon) resource
-pub static LOCATIONS_HABITATS_RESOURCE_MAP: Lazy<Mutex<HashMap<u32, u32>>> = Lazy::new(|| Mutex::new(HashMap::new()));
-
-// Map between the animation (icon resource) and the id ZT uses to reference location/habitats, this is used to lookup the id needed to add the habitat/location to an animal
-pub static LOCATIONS_HABITATS_ID_MAP: Lazy<Mutex<HashMap<String, u32>>> = Lazy::new(|| Mutex::new(HashMap::new()));
-
-// Used to ensure mod_ids don't clash, a mod will not load if an id is already in this map
-pub static MOD_ID_SET: Lazy<Mutex<HashSet<String>>> = Lazy::new(|| Mutex::new(HashSet::new()));
-
-fn add_location_or_habitat(name: &String, icon_resource_id: &String) -> anyhow::Result<()> {
-    let mut resource_binding = LOCATIONS_HABITATS_RESOURCE_MAP.lock().unwrap();
-
-    let mut id_binding = LOCATIONS_HABITATS_ID_MAP.lock().unwrap();
-
-    let string_id = add_string_to_registry(name.clone());
-
-    info!("Adding location/habitat: {} {} -> {}", name, icon_resource_id, string_id);
-
-    let icon_resource_id_cstring = CString::new(icon_resource_id.clone())
-        .with_context(|| format!("Failed to create cstring for location/habitat {} with icon_resource_id {}", name, icon_resource_id))?;
-    resource_binding.insert(string_id, icon_resource_id_cstring.into_raw() as u32);
-    id_binding.insert(name.clone(), string_id);
-
-    Ok(())
-}
-
-pub fn get_location_or_habitat_by_id(id: u32) -> Option<u32> {
-    let binding = LOCATIONS_HABITATS_RESOURCE_MAP.lock().unwrap();
-    binding.get(&id).cloned()
-}
-
-fn get_location_or_habitat_by_name(name: &String) -> Option<u32> {
-    let binding = LOCATIONS_HABITATS_ID_MAP.lock().unwrap();
-    binding.get(name).cloned()
-}
-
-// Adds a new mod id to the set, returns false if the mod_id already exists
-fn add_new_mod_id(mod_id: &str) -> bool {
-    let mut binding = MOD_ID_SET.lock().unwrap();
-    binding.insert(mod_id.to_string())
-}
-
-enum ResourceType {
-    Location,
-    Habitat,
-}
-
-impl Display for ResourceType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            ResourceType::Location => write!(f, "location"),
-            ResourceType::Habitat => write!(f, "habitat"),
-        }
-    }
-}
-
-enum ZTResourceType {
-    Animation,
-    Ani,
-    Palette,
-}
-
-impl Display for ZTResourceType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            ZTResourceType::Animation => write!(f, "animation"),
-            ZTResourceType::Palette => write!(f, "palette"),
-            ZTResourceType::Ani => write!(f, "ani"),
-        }
-    }
-}
-
-fn load_def(mod_id: &String, file_name: &String, file_map: &HashMap<String, Box<[u8]>>) -> anyhow::Result<mods::ModDefinition> {
-    info!("Loading defs {} from {}", file_name, mod_id);
-
-    let file = file_map
-        .get(file_name)
-        .with_context(|| format!("Error finding file {} in resource map for mod {}", file_name, mod_id))?;
-
-    let intermediate_string = str::from_utf8(file)
-        .with_context(|| format!("Error converting file {} to utf8 for mod {}", file_name, mod_id))?
-        .to_string();
-
-    let defs = toml::from_str::<mods::ModDefinition>(&intermediate_string).with_context(|| format!("Error parsing defs from OpenZT mod: {}", file_name))?;
-
-    info!("Loading defs: {}", defs.len());
-
-    // Habitats
-    if let Some(habitats) = defs.habitats() {
-        for (habitat_name, habitat_def) in habitats.iter() {
-            let base_resource_id = openzt_base_resource_id(mod_id, ResourceType::Habitat, habitat_name);
-            load_icon_definition(
-                &base_resource_id,
-                habitat_def,
-                file_map,
-                mod_id,
-                include_str!("../../resources/include/infoimg-habitat.ani").to_string(),
-            )?;
-            add_location_or_habitat(habitat_def.name(), &base_resource_id)?;
-        }
-    }
-
-    // Locations
-    if let Some(locations) = defs.locations() {
-        for (location_name, location_def) in locations.iter() {
-            let base_resource_id = openzt_base_resource_id(mod_id, ResourceType::Location, location_name);
-            load_icon_definition(
-                &base_resource_id,
-                location_def,
-                file_map,
-                mod_id,
-                include_str!("../../resources/include/infoimg-location.ani").to_string(),
-            )?;
-            add_location_or_habitat(location_def.name(), &base_resource_id)?;
-        }
-    }
-    Ok(defs)
-}
-
-fn load_icon_definition(
-    base_resource_id: &String,
-    icon_definition: &mods::IconDefinition,
-    file_map: &HashMap<String, Box<[u8]>>,
-    mod_id: &String,
-    base_config: String,
-) -> anyhow::Result<()> {
-    let icon_file = file_map.get(icon_definition.icon_path()).with_context(|| {
-        format!(
-            "Error loading openzt mod {}, cannot find file {} for icon_def {}",
-            mod_id,
-            icon_definition.icon_path(),
-            icon_definition.name()
-        )
-    })?;
-
-    let icon_file_palette = file_map.get(icon_definition.icon_palette_path()).with_context(|| {
-        format!(
-            "Error loading openzt mod {}, cannot find file {} for icon_def {}",
-            mod_id,
-            icon_definition.icon_palette_path(),
-            icon_definition.name()
-        )
-    })?;
-
-    let palette_file_name = openzt_full_resource_id_path(base_resource_id, ZTResourceType::Palette);
-    let palette_ztfile = ZTFile::new_raw_bytes(palette_file_name.clone(), icon_file_palette.len() as u32, icon_file_palette.clone());
-    add_ztfile(Path::new("zip::./openzt.ztd"), palette_file_name.clone(), palette_ztfile);
-
-    let mut animation = Animation::parse(icon_file);
-    animation.set_palette_filename(palette_file_name.clone());
-    let (new_animation_bytes, icon_size) = animation.write();
-
-    let new_icon_file = new_animation_bytes.into_boxed_slice();
-
-    let mut ani_cfg = Ini::new_cs();
-    ani_cfg.set_comment_symbols(&[';', '#', ':']);
-    ani_cfg.read(base_config).map_err(|s| anyhow!("Error reading ini: {}", s))?;
-
-    if ani_cfg.set("animation", "dir1", Some(openzt_full_resource_id_path(base_resource_id, ZTResourceType::Animation))).is_none() {
-        return Err(anyhow!("Error setting dir1 for ani"));
-    }
-
-    let mut write_options = WriteOptions::default();
-    write_options.space_around_delimiters = true;
-    write_options.blank_lines_between_sections = 1;
-    let new_string = ani_cfg.pretty_writes(&write_options);
-    info!("New ani: \n{}", new_string);
-    let file_size = new_string.len() as u32;
-    let file_name = openzt_full_resource_id_path(base_resource_id, ZTResourceType::Ani);
-
-    let Ok(new_c_string) = CString::new(new_string) else {
-        return Err(anyhow!(
-            "Error loading openzt mod {} when converting .ani to CString after modifying {}",
-            mod_id,
-            file_name
-        ));
-    };
-
-    let ztfile = ZTFile::new_text(file_name.clone(), file_size, new_c_string)
-        .with_context(|| format!("Error loading openzt mod {} when creating ZTFile for .ani after modifying {}", mod_id, file_name))?;
-
-    add_ztfile(Path::new("zip::./openzt.ztd"), file_name, ztfile);
-
-    let animation_file_name = openzt_full_resource_id_path(base_resource_id, ZTResourceType::Animation);
-    let animation_ztfile = ZTFile::new_raw_bytes(animation_file_name.clone(), icon_size as u32, new_icon_file);
-
-    add_ztfile(Path::new("zip::./openzt.ztd"), animation_file_name.clone(), animation_ztfile);
-
-    let palette_file_name = openzt_full_resource_id_path(base_resource_id, ZTResourceType::Palette);
-    let palette_ztfile = ZTFile::new_raw_bytes(palette_file_name.clone(), icon_file_palette.len() as u32, icon_file_palette.clone());
-    add_ztfile(Path::new("zip::./openzt.ztd"), palette_file_name, palette_ztfile);
-
-    Ok(())
-}
-
-fn openzt_base_resource_id(mod_id: &String, resource_type: ResourceType, resource_name: &String) -> String {
-    let resource_type_name = resource_type.to_string();
-    format!("openzt.mods.{}.{}.{}", mod_id, resource_type_name, resource_name)
-}
-
-fn openzt_full_resource_id_path(base_resource_id: &String, file_type: ZTResourceType) -> String {
-    format!("{}.{}", base_resource_id, file_type)
-}
-
-static RESOURCE_HANDLER_ARRAY: Lazy<Mutex<Vec<Handler>>> = Lazy::new(|| Mutex::new(Vec::new()));
-
-pub fn add_handler(handler: Handler) {
-    let mut data_mutex = RESOURCE_HANDLER_ARRAY.lock().unwrap();
-    data_mutex.push(handler);
-}
-
-fn get_handlers() -> Vec<Handler> {
-    RESOURCE_HANDLER_ARRAY.lock().unwrap().clone()
-}
